Add files via upload
Planning to make a new game that strips down on the repetitive and ugly code, as well as implements 4 new features: No program termination when entering a non-integer as a wager, introduce black and red cards, print J,K,Q,A on previous deck reveal instead of 11,12,13,14, and modify user card location on the bitmap.

This first push just has a refined unshuffled deck initialization, shuffle, and a method that can fill and rectangle on the bitmap given six arguments. It also draws the rectangle in two parts, with two different registers holding color (allows for split screen decisions).