Add files via upload
Frontend complete, with exception of only escaping on wagers when 'e' is pressed (instead of any non integer). Other than that, bust values and ace counts are tracked, cards can be red or black, deck is working, printed decks are printing J,Q,K,A not 11,12,13,14. 

Next need to prompt the user, then go to the betting and dealer protocol. Finally end screen, new hand.