Add files via upload
Fully operational front end with the exception of prompt_wager. I want to make it so non-integer inputs lead to a reprompt instead of the program crashing (unless the user answers with commas, which will be allowed). Right now I have done none of that, but everything leading to the user's first decision (hit, stand, ...) except that is done.